aider: Adjusted `projectTable.js` to directly set data in the EJS template and passed `elementsData` to the Tabulator table in `projectHome.ejs`.
# Aider chat conversation:

USER: we already have projectTable.js

ASSISTANT: Given that `projectTable.js` is already present in the repository and is responsible for initializing the Tabulator table, we need to adjust it to work with the new JSON data structure that is output by the `analyze_ifc.py` script and passed to the `projectHome.ejs` view.

Here are the diffs for the changes to `public/js/projectTable.js`:

```diff
--- public/js/projectTable.js
+++ public/js/projectTable.js
@@ ... @@
 var table = new Tabulator("#elements-table", {
     height: "auto",
     layout: "fitColumns",
-    ajaxURL: "/api/projects/" + projectId + "/elements", // URL to fetch data
+    // Removed ajaxURL as data will be passed directly to the table in the EJS template
     columns: [
         {title: "GUID", field: "guid"},
         {title: "Name", field: "name"},
         {title: "Type", field: "type"},
         {title: "IfcClass", field: "ifc_class"},
         {title: "BuildingStorey", field: "building_storey"},
         {title: "IsLoadbearing", field: "is_loadbearing"},
         {title: "IsExternal", field: "is_external"},
         {title: "Volume", field: "volume"},
         {title: "Surface", field: "surface"},
         {title: "Material", field: "material"},
         // ... rest of the columns ...
     ],
     // Additional Tabulator options for sorting, filtering, and collapsing will be added here
     // ...
+    // Data will be set directly in the EJS template, so no need to fetch it via ajax
+    data: elementsData, // elementsData will be passed from EJS template
 });
```

And here are the diffs for the changes to `views/projectHome.ejs` to pass the `elementsData` directly to the Tabulator table:

```diff
--- views/projectHome.ejs
+++ views/projectHome.ejs
@@ ... @@
 <script src="/js/uploadIfc.js"></script>
 <script>
     var projectId = "<%= project._id %>";
+    // Pass elementsData from the server to the Tabulator table
+    var elementsData = <%- JSON.stringify(elementsData) %>;
 </script>
 <script src="/js/projectTable.js"></script>
 </body>
 </html>
```

These changes will ensure that the `elementsData` received from the server is passed directly to the Tabulator table in `projectTable.js`. The `ajaxURL` property is removed because the data is no longer being fetched asynchronously; it is provided by the server when the page is rendered. The `data` property is added to the Tabulator initialization to set the table data directly. The `elementsData` is passed to the `projectTable.js` script by setting a JavaScript variable in the `projectHome.ejs` template.